+++ tests/test_0200_configure.py
@@ -19,32 +19,32 @@ def common_asserts(result, substring: Optional[list] = None, exit_code: int = 0)
 
 
 def test_configure_tenant_via_flags_no_alias(runner, cli):
-    result = runner.invoke(cli, 'configure tenant -t pybritivetest1.dev -f yaml'.split(' '))
+    result = runner.invoke(cli, ['configure', 'tenant', '-t', 'pybritivetest1.dev', '-f', 'yaml'])
     common_asserts(result, substring='[tenant-pybritivetest1.dev]')
 
 
 def test_configure_tenant_via_flags_no_alias_no_format(runner, cli):
-    result = runner.invoke(cli, 'configure tenant -t pybritivetest2.dev'.split(' '))
+    result = runner.invoke(cli, ['configure', 'tenant', '-t', 'pybritivetest2.dev'])
     common_asserts(result, substring='[tenant-pybritivetest2.dev]')
 
 
 def test_configure_tenant_via_flags_yes_alias(runner, cli):
-    result = runner.invoke(cli, 'configure tenant -t pybritivetest1.dev -f yaml -a testalias1'.split(' '))
+    result = runner.invoke(cli, ['configure', 'tenant', '-t', 'pybritivetest1.dev', '-f', 'yaml', '-a', 'testalias1'])
     common_asserts(result, substring='[tenant-testalias1]')
 
 
 def test_configure_tenant_via_flags_yes_alias_no_format(runner, cli):
-    result = runner.invoke(cli, 'configure tenant -t pybritivetest2.dev -a testalias2'.split(' '))
+    result = runner.invoke(cli, ['configure', 'tenant', '-t', 'pybritivetest2.dev', '-a', 'testalias2'])
     common_asserts(result, substring='[tenant-testalias2]')
 
 
 def test_configure_tenant_via_prompt_no_alias(runner, cli):
-    result = runner.invoke(cli, 'configure tenant'.split(' '), input='pybritivetest3.dev\n\njson\n')
+    result = runner.invoke(cli, ['configure', 'tenant'], input='pybritivetest3.dev\n\njson\n')
     common_asserts(result, substring='[tenant-pybritivetest3.dev]')
 
 
 def test_configure_tenant_via_prompt_no_alias_no_format(runner, cli):
-    result = runner.invoke(cli, 'configure tenant'.split(' '), input='pybritivetest4.dev\n\n\n')
+    result = runner.invoke(cli, ['configure', 'tenant'], input='pybritivetest4.dev\n\n\n')
     common_asserts(result, substring='[tenant-pybritivetest4.dev]')
 
 
@@ -59,14 +59,16 @@ def test_configure_tenant_via_prompt_yes_alias_no_format(runner, cli):
 
 
 def test_configure_global_via_flags_file_backend(runner, cli):
-    result = runner.invoke(cli, 'configure global -t pybritivetest1.dev -f table -b file'.split(' '))
+    result = runner.invoke(cli, ['configure', 'global', '-t', 'pybritivetest1.dev', '-f', 'table', '-b', 'file'])
     common_asserts(
         result, substring=['default_tenant=pybritivetest1.dev', 'output_format=table', 'credential_backend=file']
     )
 
 
 def test_configure_global_via_flags_encrypted_file_backend(runner, cli):
-    result = runner.invoke(cli, 'configure global -t pybritivetest2.dev -f yaml -b encrypted-file'.split(' '))
+    result = runner.invoke(
+        cli, ['configure', 'global', '-t', 'pybritivetest2.dev', '-f', 'yaml', '-b', 'encrypted-file']
+    )
     common_asserts(
         result,
         substring=['default_tenant=pybritivetest2.dev', 'output_format=yaml', 'credential_backend=encrypted-file'],
@@ -74,14 +76,14 @@ def test_configure_global_via_flags_encrypted_file_backend(runner, cli):
 
 
 def test_configure_global_via_prompt_file_backend(runner, cli):
-    result = runner.invoke(cli, 'configure global'.split(' '), input='pybritivetest1.dev\ntable-pretty\nfile\n')
+    result = runner.invoke(cli, ['configure', 'global'], input='pybritivetest1.dev\ntable-pretty\nfile\n')
     common_asserts(
         result, substring=['default_tenant=pybritivetest1.dev', 'output_format=table-pretty', 'credential_backend=file']
     )
 
 
 def test_configure_global_via_prompt_encrypted_file_backend(runner, cli):
-    result = runner.invoke(cli, 'configure global'.split(' '), input='pybritivetest2.dev\n\nencrypted-file\n')
+    result = runner.invoke(cli, ['configure', 'global'], input='pybritivetest2.dev\n\nencrypted-file\n')
     common_asserts(
         result,
         substring=['default_tenant=pybritivetest2.dev', 'output_format=json', 'credential_backend=encrypted-file'],
@@ -89,23 +91,23 @@ def test_configure_global_via_prompt_encrypted_file_backend(runner, cli):
 
 
 def test_configure_global_with_invalid_format(runner, cli):
-    result = runner.invoke(cli, 'configure global -f error -P'.split(' '))
+    result = runner.invoke(cli, ['configure', 'global', '-f', 'error', '-P'])
     assert "Invalid value for '--format' / '-f'" in result.output
 
 
 def test_configure_global_with_invalid_tenant(runner, cli):
-    result = runner.invoke(cli, 'configure global -t incorrect'.split(' '))
+    result = runner.invoke(cli, ['configure', 'global', '-t', 'incorrect'])
     assert 'Invalid global field default_tenant value incorrect provided. Tenant not found.' in result.output
 
 
 def test_configure_update_global_invalid_data(runner, cli):
-    result = runner.invoke(cli, 'configure update global default_tenant incorrect'.split(' '))
+    result = runner.invoke(cli, ['configure', 'update', 'global', 'default_tenant', 'incorrect'])
     assert result.exit_code == 1
     assert 'Invalid global field default_tenant value incorrect provided. Tenant not found.' in result.output
 
 
 def test_configure_update_invalid_section(runner, cli):
-    result = runner.invoke(cli, 'configure update test default_tenant incorrect'.split(' '))
+    result = runner.invoke(cli, ['configure', 'update', 'test', 'default_tenant', 'incorrect'])
     assert result.exit_code == 1
     assert 'Cannot save config file due to invalid data provided.' in result.output
 
@@ -123,5 +125,5 @@ def test_configure_update_global_correct_data(runner, cli):
 
 
 def test_configure_update_aws_data(runner, cli):
-    result = runner.invoke(cli, 'configure update aws default_checkout_mode integrate'.split(' '))
+    result = runner.invoke(cli, ['configure', 'update', 'aws', 'default_checkout_mode', 'integrate'])
     common_asserts(result, substring=['aws', 'default_checkout_mode=integrate'])