+++ tests/test_0500_api.py
@@ -7,5 +7,5 @@ def common_asserts(result, substring=None, exit_code=0):
 
 
 def test_api(runner, cli):
-    result = runner.invoke(cli, 'api identity_management.users.list'.split(' '))
+    result = runner.invoke(cli, ['api', 'identity_management.users.list'])
     common_asserts(result, ['userId', 'status', 'email', 'identityProvider'])